NEX-975: Create a workflow to package helm charts (#21)
The process to create a package, update the index, and submit to a PR is complicated/manual. This is because of the created tag being updated in the index.yaml for every package even though you may have only touched one. This makes that process automated in the PR.

The github actions bot will see your change to the helm charts source, create the new package, update the index, and push a commit to your PR which then you can merge.

Here is an example of it being triggered: https://github.com/Bluescape/helm-charts/runs/2291585413?check_suite_focus=true
Here is the example patch which it applied: 0fe557e